Added ViewWindow and showView as general functions, used from tests now
