+++ datavis/views/__init__.py
@@ -21,3 +21,5 @@ from .picker_view import (PickerView, SHAPE_RECT, SHAPE_CIRCLE, SHAPE_SEGMENT,
                           DEFAULT_MODE, FILAMENT_MODE, SHAPE_CENTER)
 
 from ._paging_view import PagingView
+
+from ._utils import showView, ViewWindow
